requirements.txt
annotated-types==0.7.0
anyio==4.8.0
authlib==1.3.1 # Added for potential OAuth utilities (optional but good)
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
fastapi-cli==0.0.7
graphql-core>=3.2.1,<4.0.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
orjson==3.10.15
packaging==24.2
prisma>=0.15.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0 # Added for JWT validation
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette>=0.40.0,<0.47.0
starlette-session==0.4.3 # Added for session management
strawberry-graphql==0.262.4
StrEnum==0.4.15
tomlkit==0.13.2
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
